backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
greenlet>=3.0.0
python-dotenv>=1.0.1
tweepy>=4.14.0
requests>=2.31.0
requests-oauthlib>=1.3.1
openai>=1.57.0
anthropic>=0.39.0
apscheduler>=3.10.4
python-multipart>=0.0.17
pydantic>=2.10.0
pydantic-settings>=2.6.0
PyPDF2>=3.0.1
python-docx>=1.1.0
markdown>=3.7
beautifulsoup4>=4.12.3
aiofiles>=24.1.0
httpx>=0.28.0
numpy>=1.26.0
matplotlib>=3.8.0
gymnasium>=0.29.0
stable-baselines3>=2.3.0
PyJWT>=2.8.0
cryptography>=42.0.0
